Add ability to deploy Capsules for a given server
This introduces the ability to auto-deploy a Capsule configured
and running against any server. Further, this introduces the use
of Ansible as a tool to do the orchestrating between Server and Capsule
and is thus required on the host doing the deployments.
The boxes.yaml paradigm can be used to spin up multiple external
Caspules for a given server.