Add data import tasks
These tasks will run a job for each PV to move data from an existing
location to the PV using rsync.

These tasks are included into the existing foreman.yml playbook
for the start tag if the data_import variable is set to `true`.

An additional playbook is also provided to run the data import
jobs outside of the foreman application deployment process.